requirements.txt
pandas
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
